Move ssh auth into same package as credentials
