+++ credentials/ssh_test.go
@@ -1,4 +1,4 @@
-package lfs
+package credentials
 
 import (
 	"path/filepath"
